+++ MP.Core/Models/MediaFile.cs
@@ -22,6 +22,6 @@ namespace MP.Core.Models
         public long BytesPerSecond { get; set; }
         public DateTime LastProcessingUpdate { get; set; }
         public string ProcessedFormat { get; set; }
-        public long BytesPerSecondPerPixel { get; set; }
+        public double BytesPerSecondPerPixel { get; set; }
     }
 }
